feat: add build target field to routes feat: move towards single App.js file fix: header component interaction with build target
